+++ clnta.php
@@ -1,4 +1,6 @@
-
+<?php
+require("panel_new.php");
+?>
 <?php
 // require('demo_a.php');
 $conn = mysqli_connect('localhost','root','','demo');
@@ -35,7 +37,6 @@ if(isset($_POST['register'])){
 if(isset($_GET['delete'])){
    $id = $_GET['delete'];
    mysqli_query($conn, "DELETE FROM clnt_reg WHERE id = $id");
-   header('location:clnta.php');
 };
 
 ?>
@@ -119,7 +120,7 @@ if(isset($message)){
                                             <td><?php echo $row['Phn_no'];?></td>
             <td>
                <!-- <a href="admin_update.php?edit=<?php echo $row['Username']; ?>" class="btn"> <i class="fas fa-edit"></i> edit </a> -->
-               <a href="clnta.php?delete=<?php echo $row['id']; ?>" class="btn"> <i class="fas fa-trash"></i> delete </a>
+               <a href="clnta.php?delete=<?php echo $row['id']; ?>" class="btn"  onclick="return confirm('Are you sure you want to Delete');"> <i class="fas fa-trash"></i> delete </a>
             </td>
          </tr>
       <?php } ?>
