+++ servicebooka.php
@@ -1,6 +1,7 @@
-
+<?php 
+require("panel_new.php");
+?>
 <?php
-// require('demo_a.php');
 $conn = mysqli_connect('localhost','root','','demo');
 
 
@@ -31,7 +32,6 @@ if(isset($_POST['add_product'])){
 if(isset($_GET['delete'])){
     $id = $_GET['delete'];
     mysqli_query($conn, "DELETE FROM service_book WHERE id = $id");
-    header('location:servicebooka.php');
  };
  
 ?>
@@ -105,7 +105,7 @@ if(isset($message)){
                                             <td><?php echo $row['dt_book'];?> &nbsp; </td>
             <td>
                <!-- <a href="admin_update.php?edit=<?php echo $row['Username']; ?>" class="btn"> <i class="fas fa-edit"></i> edit </a> -->
-               <a href="servicebooka.php?delete=<?php echo $row['id']; ?>" class="btn"> <i class="fas fa-trash"></i> delete </a>
+               <a href="servicebooka.php?delete=<?php echo $row['id']; ?>" class="btn"  onclick="return confirm('Are you sure you want to Delete');"> <i class="fas fa-trash"></i> delete </a>
             </td>
          </tr>
       <?php } ?>
